Move ImageLabel from gui.widgets to gui.image
It is explicitly designed for being used as base by the classes in
gui.image and should therefore also be defined there.
